New code for motion detection which doesn't have to calculate all contours and look through these, this also eliminates the need for dilation and blurring, which results in an easier processing process.
Now all it does is convert image to greyscale, takes the absolute difference between the image and the previous image, and then thresholds the image to get a matrix of 0's and 1's which it then takes the sum of. This sum then determines whether or not it is considered movement.